[WIP] Stub data with default props instead of constructor
The advantage of this approach is that you can start thinking about
what props will be used in the render method. Then, when we're ready,
we can just remove the default props and pass in the real props. The
advantage to this approach is that you won't have to change the render
method when this happens; just remove the defaults and you're good to
go!