[F] Add stub for browse-featured projects
[FEATURE #108400756] This commit adds a new svg to the icon stack;
Stub component has also been built for the "Featured Projects"
section shown in the comps; This section is pulling in projects from
the getFeatureProjects method, but could easily be filtered with other
methods. It will likely require a parameter to limit the max length of
the results. Styles are also necessary for hover behaviors on "add"
buttons.